refactor: Extract config commands to a separate file
